requirement.txt
langchain-openai
langchain-community>=0.2.11
langchain-core
tavily-python
streamlit
python-dotenv
